Conflicts:

	app/models/plugin.rb
	config/init.rb

+++ app/models/plugin.rb
@@ -1,20 +1,23 @@
-class Plugin < DataMapper::Base
-  property :url, :string, :length => 255
-  property :path, :string, :length => 255
-  property :name, :string
-  property :version, :string
-  property :author_name, :string
-  property :author_email, :string
-  property :author_homepage, :string
-  property :homepage, :string, :length => 255
-  property :about, :string, :length => 255
-  property :active, :boolean
+class Plugin
+  include DataMapper::Resource
 
-  before_create :download
-  after_create :install
-  after_create :set_create_activity
-  after_update :set_update_activity
-  after_destroy :remove
+  property :id, Integer, :key => true  
+  property :url, String, :length => 255
+  property :path, String, :length => 255
+  property :name, String
+  property :version, String
+  property :author_name, String
+  property :author_email, String
+  property :author_homepage, String
+  property :homepage, String, :length => 255
+  property :about, String, :length => 255
+  property :active, Boolean
+
+  before :save, :download
+  after :save, :install
+  after :save, :set_create_activity
+  after :save, :set_update_activity
+  after :destroy, :remove
 
   class << self
     @@loaded = []
@@ -23,34 +26,38 @@ class Plugin < DataMapper::Base
   ##
   # This grabs the plugin using its url, unpacks it, and loads the metadata for it
   def download
-    # Load the manifest yaml
-    manifest = YAML::load(Net::HTTP.get(URI.parse(url)))
-    # Grab metadata from manifest
-    self.name = manifest["name"]
-    self.author = manifest["author"]
-    self.version = manifest["version"]
-    self.homepage = manifest["homepage"]
-    self.about = manifest["about"]
-    # Build the path
-    self.path = File.join(Merb.root, "app", "plugins", self.name)
-    # Remove any existing plugin at the path
-    FileUtils.rm_rf(self.path)
-    FileUtils.mkdir_p(self.path)
-    # Download the package and untgz
-    package_url = File.join(url.split('/').slice(0..-2).join('/'), manifest["package"])
-    package = Net::HTTP.get(URI.parse(package_url))
-    Archive::Tar::Minitar.unpack(Zlib::GzipReader.new(StringIO.new(package)), self.path)
-    # Unpack any gems downloaded
-    unpack_gems(Dir.glob(File.join(self.path, "gems", "*.gem")).collect { |p| p.split("/").last })
+    if new_record?
+      # Load the manifest yaml
+      manifest = YAML::load(Net::HTTP.get(URI.parse(url)))
+      # Grab metadata from manifest
+      self.name = manifest["name"]
+      self.author = manifest["author"]
+      self.version = manifest["version"]
+      self.homepage = manifest["homepage"]
+      self.about = manifest["about"]
+      # Build the path
+      self.path = File.join(Merb.root, "app", "plugins", self.name)
+      # Remove any existing plugin at the path
+      FileUtils.rm_rf(self.path)
+      FileUtils.mkdir_p(self.path)
+      # Download the package and untgz
+      package_url = File.join(url.split('/').slice(0..-2).join('/'), manifest["package"])
+      package = Net::HTTP.get(URI.parse(package_url))
+      Archive::Tar::Minitar.unpack(Zlib::GzipReader.new(StringIO.new(package)), self.path)
+      # Unpack any gems downloaded
+      unpack_gems(Dir.glob(File.join(self.path, "gems", "*.gem")).collect { |p| p.split("/").last })
+    end
   end
 
   ##
   # This loads and installs the plugin
   def install
-    # Load the plugin
-    self.load
-    # Also, if there is an "install.rb" script present, run that to setup anything the plugin needs (database tables etc)
-    require File.join(self.path, "install.rb") if File.exists?(File.join(self.path, "install.rb"))
+    if new_record?
+      # Load the plugin
+      self.load
+      # Also, if there is an "install.rb" script present, run that to setup anything the plugin needs (database tables etc)
+      require File.join(self.path, "install.rb") if File.exists?(File.join(self.path, "install.rb"))
+    end
   rescue Exception => err
     # Catch the error, delete the plugin, and raise an error again
     self.destroy!
@@ -60,17 +67,21 @@ class Plugin < DataMapper::Base
   ##
   # This adds the activity to show a plugin has been installed
   def set_create_activity
-    a = Activity.new
-    a.message = "Plugin \"#{self.name}\" installed"
-    a.save
+    if new_record?
+      a = Activity.new
+      a.message = "Plugin \"#{self.name}\" installed"
+      a.save
+    end
   end
 
   ##
   # This adds the activity to show a plugin has been updated
   def set_update_activity
-    a = Activity.new
-    a.message = "Plugin \"#{self.name}\" #{self.active ? 'activated' : 'de-activated'}"
-    a.save
+    unless new_record?
+      a = Activity.new
+      a.message = "Plugin \"#{self.name}\" #{self.active ? 'activated' : 'de-activated'}"
+      a.save
+    end
   end
 
   ##
